+++ cpp/arduino/Server.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <Stream.h>
+
+class Server : public Print {};